% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cert.R
\name{addTargetDrugDataFrame}
\alias{addTargetDrugDataFrame}
\title{Add a target drug information into existing TargetDrugDataFrame}
\usage{
addTargetDrugDataFrame(origin, name, class, code)
}
\arguments{
\item{origin}{Name of existing TargetDrugDataFrame variable}

\item{name}{Name of drug}

\item{class}{Classification of drug}

\item{code}{Drug code in a specific classification}
}
\value{
A data frame
}
\description{
Add a target drug information into existing TargetDrugDataFrame
}

