% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cert.R
\name{createLabtestDataFrame}
\alias{createLabtestDataFrame}
\title{Create data frame for set a target labtest}
\usage{
createLabtestDataFrame(id, name, type)
}
\value{
nothing
}
\description{
Create data frame for set a target labtest
}
\details{
initial testing
}

