% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cert.R
\name{createTargetDrugDataFrame}
\alias{createTargetDrugDataFrame}
\title{Create data frame for a target drug information}
\usage{
createTargetDrugDataFrame(name, class, code)
}
\arguments{
\item{name}{Name of drug}

\item{class}{Classification of drug}

\item{code}{Drug code in a specific classification}
}
\value{
A data frame
}
\description{
Create data frame for a target drug information
}

