% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cert.R
\name{generateCertDataSet}
\alias{generateCertDataSet}
\title{Generate Data set for Cert}
\usage{
generateCertDataSet(connectionDetails, drug_list = NA, labtest_list = NA,
  date_from = "2001-01-01", date_to = "2010-03-31")
}
\arguments{
\item{connectionDetails}{connectionDetails information
generated by DatabaseConnector::createConnectionDetails function}

\item{drug_list}{Drug information list
generated by createTargetDrugDataFrame function}

\item{labtest_list}{Laboratory test information list
generated by createLabtestDataFrame function}

\item{date_from}{Begin date of history
By default 2001-01-01}

\item{date_to}{End date of history
By default 2010-03-31}
}
\value{
Nothing
}
\description{
By using drug and lab-test list, user can generate dataset for Cert
}

