% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cert.R
\name{getCertDemographics}
\alias{getCertDemographics}
\title{Get demographics data set of Cert}
\usage{
getCertDemographics(connectionDetails)
}
\arguments{
\item{connectionDetails}{connectionDetails information
generated by DatabaseConnector::createConnectionDetails function}
}
\value{
A data frame from demographics table
generated by Cert algorithm
}
\description{
Demographics data set of Cert algorithm
}

