% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cert.R
\name{getCertResultDataSet}
\alias{getCertResultDataSet}
\title{Get result data set of Cert}
\usage{
getCertResultDataSet(connectionDetails)
}
\arguments{
\item{connectionDetails}{connectionDetails information
generated by DatabaseConnector::createConnectionDetails function}
}
\value{
A data frame from summary table
generated by Cert algorithm
}
\description{
Summary data set of Cert algorithm
}

