% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cert.R
\name{getDataForMcNemarTest}
\alias{getDataForMcNemarTest}
\title{Get Cert result dataset for McNemar's test}
\usage{
getDataForMcNemarTest(connectionDetails)
}
\value{
a data frame from summary table
}
\description{
Get Cert result dataset for McNemar's test
}
\details{
initial testing
}

