% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cert.R
\name{getDataForPairedTTest}
\alias{getDataForPairedTTest}
\title{Get Cert result dataset for paired t-test}
\usage{
getDataForPairedTTest(connectionDetails)
}
\value{
a data frame from summary table
}
\description{
Get Cert result dataset for paired t-test
}
\details{
initial testing
}

