% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cert.R
\name{queryTest}
\alias{queryTest}
\title{Test query database}
\usage{
queryTest(connectionDetails)
}
\arguments{
\item{connectionDetails}{connectionDetails information}
}
\value{
a data frame
}
\description{
Test query database
}
\details{
for testing
}

