% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cert.R
\name{runMcNemarTest}
\alias{runMcNemarTest}
\title{Run McNemar's Test}
\usage{
runMcNemarTest(connectionDetails)
}
\arguments{
\item{connectionDetails}{connectionDetails information
generated by DatabaseConnector::createConnectionDetails function}
}
\value{
a data frame
outcome of McNemar's test
}
\description{
Extract data set and run McNemar's test
}

