% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cert.R
\name{runPairedTTest}
\alias{runPairedTTest}
\title{Run paired t-test}
\usage{
runPairedTTest(connectionDetails)
}
\arguments{
\item{connectionDetails}{connectionDetails information
generated by DatabaseConnector::createConnectionDetails function}
}
\value{
A data frame
outcome of paired t-test
}
\description{
Extract data set and run paired t-test
}

