% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cert.R
\name{addTargetDrugDataFrame}
\alias{addTargetDrugDataFrame}
\title{Add into TargetDrugDataFrame}
\usage{
addTargetDrugDataFrame(origin, name, class, code)
}
\value{
nothing
}
\description{
Add into TargetDrugDataFrame
}
\details{
initial testing
}

