% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cert.R
\name{createTargetDrugDataFrame}
\alias{createTargetDrugDataFrame}
\title{Create data frame for set a target drug}
\usage{
createTargetDrugDataFrame(name, class, code)
}
\value{
nothing
}
\description{
Create data frame for set a target drug
}
\details{
initial testing
}

