% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cert.R
\name{generateCertDataSet}
\alias{generateCertDataSet}
\title{Generate Cert dataset}
\usage{
generateCertDataSet(connectionDetails, drug_list = NA, labtest_list = NA,
  date_from = "2001-01-01", date_to = "2010-03-31")
}
\value{
nothing
}
\description{
Generate Cert dataset
}
\details{
initial testing
}

