% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cert.R
\name{getCertDemographics}
\alias{getCertDemographics}
\title{Get Cert demographics dataset}
\usage{
getCertDemographics(connectionDetails)
}
\value{
a data frame from demographics table
}
\description{
Get Cert demographics dataset
}
\details{
initial testing
}

