% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cert.R
\name{getCertResultDataSet}
\alias{getCertResultDataSet}
\title{Get Cert result dataset}
\usage{
getCertResultDataSet(connectionDetails)
}
\value{
a data frame from summary table
}
\description{
Get Cert result dataset
}
\details{
initial testing
}

