% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cert.R
\name{queryTest}
\alias{queryTest}
\title{Test query database}
\usage{
queryTest(connectionDetails)
}
\arguments{
\item{connectionDetails}{connectionDetails information
generated by DatabaseConnector::createConnectionDetails function}
}
\value{
A data frame
}
\description{
Test query for check connection with database
}

